moved fetch logic to useArchiveView()
